requirements.txt
minimalmodbus
python-periphery
